part4 steps x-10 blog updates

+++ controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const notes = await Blog
@@ -9,10 +11,22 @@ blogsRouter.get('/', async (request, response) => {
   response.json(notes)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
 
-  const user = await User.findById(body.userId)
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(401).json({ error: 'user not found' })
+  }
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: 'title or url is missing' })
@@ -33,27 +47,52 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedNote)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog){
+    return response.status(401).json({ error: 'Cant find blog' })
+  }
+
+  const user = request.user
+  if (!request.user) {
+    return response.status(401).json({ error: 'user not found' })
+  }
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).json({ error: 'no access' })
+  }
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
-  try {
-    const body = request.body
+blogsRouter.put('/:id', middleware.userExtractor, async (request, response) => {
+  const blogById = await Blog.findById(request.params.id)
+  const body = request.body
 
-    const blog = {
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes
-    }
+  const blog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes
+  }
 
-    const updated = await Blog
-      .findByIdAndUpdate(request.params.id, blog, { new: true })
+  if (!blog) {
+    return response.status(401).json({ error: 'Can\'t find blog' })
+  }
+
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'User not found' })
+  }
+
+  if (blogById.user.toString() === user._id.toString()) {
+    const updated = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     response.json(updated)
-  } catch (error) {
-    error => next(error)
+  } else {
+    response.status(403).json({ error: 'No access' })
   }
 })
